requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.1
django-appconf==1.0.4
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-image-cropping==1.5.0
django-phonenumber-field==5.0.0
django-rest-passwordreset==1.1.0
django-smtp-ssl==1.0
django-smuggler==1.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
easy-thumbnails==2.7
fcm-django==0.3.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
phonenumberslite==8.12.10
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
simplejson==3.17.2
six==1.13.0
sqlparse==0.3.1
sslcommerz-python==0.0.6
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
zope.interface==5.1.2

